fix :error in eslint script
