+++ src/apps/front-office/utils/types.ts
@@ -55,8 +55,40 @@ export type CreatedAt = {
   offset: number;
   humanTime: string;
   text: string;
+  date?: string;
+  time?: string;
 };
 
+export interface UpdatedAt {
+  format: string;
+  timestamp: number;
+  offset: number;
+  humanTime: string;
+  text: string;
+  date: string;
+}
+
+export interface CreatedBy {
+  id: number;
+  isCustomer: boolean;
+  name: string;
+  phoneNumber: string;
+  email: string;
+  userType: string;
+  totalWishlist: number;
+  totalCompare: number;
+}
+
+export interface UpdatedBy {
+  id: number;
+  isCustomer: boolean;
+  name: string;
+  phoneNumber: string;
+  email: string;
+  userType: string;
+  totalWishlist: number;
+  totalCompare: number;
+}
 export type Discount = {
   percentage: number;
   amount: number;
@@ -94,17 +126,56 @@ export type ProductCategory = {
   name: string;
 };
 
+export interface Category3 {
+  id: number;
+  isActive: boolean;
+  name: LocaleValue[];
+  image?: Image;
+  slug: string;
+}
+
+export type Product1 = {
+  id: number;
+  category: Category3;
+  createdAt: CreatedAt;
+  description: LocaleValue[];
+  discount: Discount;
+  hasDiscount: boolean;
+  images: Image[];
+  inCart: boolean;
+  inCompare: boolean;
+  inStock: boolean;
+  inWishlist: boolean;
+  isActive: boolean;
+  isLowStock: boolean;
+  name: LocaleValue[];
+  price: number;
+  purchase: {
+    minQuantity: number;
+  };
+  salePrice: number;
+  shortDescription: LocaleValue[];
+  slug: string;
+  sortOrder: number;
+  type: string;
+  rating: number;
+};
 export interface Category2 {
   id: number;
-  name: string;
-  items: number;
-  imgSrc: string;
+  isActive: boolean;
+  name: LocaleValue[];
+  image?: Image;
+  slug: string;
+  totalProducts: number;
+  products: Product1[];
 }
 
 export interface SliderData {
   categories: Category2[];
   sectionTitle: {
-    localeCode: string;
-    value: string;
+    userId: number;
+    id: number;
+    title: string;
+    completed: boolean;
   };
 }
